refactor(projects): refactor according to sourcery-ai
