+++ src/pages/projects.module.css
@@ -70,5 +70,5 @@
   font-size: 1.5rem;
   font-weight: bold;
   margin-bottom: 0.5rem;
-  color: var(--text-color);
+  color: var(--ifm-font-color-base);
 }
